Miscellaneous changes:
1. Archival now attempts to change permissions on only files owned by
the user running the pipeline (otherwise errors are raised).
2. CombineVariants now bgzips and indexes the generated VCF, rather than
handling it in an unrelated method.
3. Rewrote some of the logic/error checking for fix_phased_vcf(),
including getting GATK annotations from the analysisReady VCF directly,
instead of a "deannotated" VCF which is never created in the pipeline.
4. Pipeline is now tolerant of custom capture samples without an X or Y
targets.  Adjusted the rsync command to accommodate this.
5. Fixed obtaining X & Y target BEDs to make sure they're obtained in
the proper order without relying on non-standard MySQL feature.
6. Fixed prerequisites for both RunCvgMetrics and AlignmentMetrics to
PrintReads (BQSR).