Further changes to organization of pipeline:
1. Deleted deprecated CopyBam step.
2. Deleted obvious/unnecessary comments.
3. Removed recording commands to self.script in indivudal tasks; moved
it to parent class.
4. Set the number of cores to use for commands in classes to n_cpu
instead of being hardcoded.
5. Added abstract method _run_post_success() in order to perform any
cleanup operations (e.g. deleting a previously needed BAM) only after
setting the task as a success in the database.
6. Removed adding a deprecated AnnoDBID annotation.
7. Moved some variable declarations from __init__ to pre_shell_commands
so as to not have to set them every time the class is instantiated.
8. Implemented tarballing scripts, logs, and .txt in one file, coverage
bins in another, and GQ bins in a third.
9. Ensured obtaining X & Y BED files in the proper order.
10. Removed commit() after updating a QC metric in favor of only
committing at the end to avoid integrity errors if a crash occurs during
this step and it's restarted.
11. Every task sets the user's umask so as to preserve group write
permissions on all created files.
12. sample_name, capture_kit_bed, sample_type, and scratch are now
optional parameters, and are obtained from SeqDB if not specified.