Extensive changes.
1. Commented out (temporarily) data deletions.  Suggested to reintroduce
upon further inclusion of error checking to ensure only happens when
intended.
2. Only allow max of 3 alternate alleles in an active window to avoid
GATK bug (and should only go above this in areas of very questionable
quality).
3. Switched java ParallelGCThreads to 1 so as not to invalidate the
resources requested on the cluster.
4. Changed get_productionvcf to get the VCF on the basis of the prepID
instead of pseudo_prepid, as pseudo_prepid is not always set in the
given table and we already query for it.
5. Fixed logic to get the production VCF and ensure it exists. NB. it
doesn't exist if the previous pipeline wasn't actually run on the
sample.
6. Fixed defining the output file names for DP bin splitting.
7. Changed to always retain chromosome MT DP bins (even though they're
not in the defined regions of interest).
8. GenotypeConcordance raises an error message if no production VCF
exists.
9. Fixed bug getting FILTER field in GenotypeConcordance.subset_vcf.
10. UpdateSeqdbMetrics now requires SplitAndSubsetDPBins instead of
CvgBinning.